+++ dependencies/predis/predis/src/Transaction/AbortedMultiExecException.php
@@ -27,9 +27,10 @@ class AbortedMultiExecException extends PredisException
      * @param string    $message     Error message.
      * @param int       $code        Error code.
      */
-    public function __construct(MultiExec $transaction, $message, $code = null)
+    public function __construct(MultiExec $transaction, $message, $code = 0)
     {
         parent::__construct($message, is_null($code) ? 0 : $code);
+
         $this->transaction = $transaction;
     }