+++ includes/object-cache.php
@@ -1,13 +1,14 @@
 <?php
 /**
  * Plugin Name: Redis Object Cache Drop-In
- * Plugin URI: http://wordpress.org/plugins/redis-cache/
- * Description: A persistent object cache backend powered by Redis. Supports Predis, PhpRedis, Credis, HHVM, replication, clustering and WP-CLI.
- * Version: 2.0.26
+ * Plugin URI: https://wordpress.org/plugins/redis-cache/
+ * Description: A persistent object cache backend powered by Redis. Supports Predis, PhpRedis, Relay, replication, sentinels, clustering and WP-CLI.
+ * Version: 2.1.0
  * Author: Till Krüss
  * Author URI: https://objectcache.pro
  * License: GPLv3
  * License URI: http://www.gnu.org/licenses/gpl-3.0.html
+ * Requires PHP: 7.0
  *
  * @package Rhubarb\RedisCache
  */
@@ -461,6 +462,13 @@ class WP_Object_Cache {
     ];
 
     /**
+     * List of groups and their types.
+     *
+     * @var array
+     */
+    public $group_type = [];
+
+    /**
      * Prefix used for global groups.
      *
      * @var string
@@ -526,6 +534,8 @@ class WP_Object_Cache {
             $this->unflushable_groups = array_map( [ $this, 'sanitize_key_part' ], WP_REDIS_UNFLUSHABLE_GROUPS );
         }
 
+        $this->cache_group_types();
+
         if ( defined( 'WP_REDIS_TRACE' ) && WP_REDIS_TRACE ) {
             $this->trace_enabled = true;
         }
@@ -554,7 +564,7 @@ class WP_Object_Cache {
             }
 
             if ( defined( 'WP_REDIS_CLUSTER' ) ) {
-                $connectionID = current( array_values( WP_REDIS_CLUSTER ) );
+                $connectionID = current( $this->build_cluster_connection_array() );
 
                 $this->diagnostics[ 'ping' ] = ($client === 'predis')
                     ? $this->redis->getClientFor( $connectionID )->ping()
@@ -574,6 +584,25 @@ class WP_Object_Cache {
         if ( function_exists( 'is_multisite' ) ) {
             $this->global_prefix = is_multisite() ? '' : $table_prefix;
             $this->blog_prefix = is_multisite() ? $blog_id : $table_prefix;
+        }
+    }
+
+    /**
+     * Set group type array
+     *
+     * @return void
+     */
+    protected function cache_group_types() {
+        foreach ( $this->global_groups as $group ) {
+            $this->group_type[ $group ] = 'global';
+        }
+
+        foreach ( $this->unflushable_groups as $group ) {
+            $this->group_type[ $group ] = 'unflushable';
+        }
+
+        foreach ( $this->ignored_groups as $group ) {
+            $this->group_type[ $group ] = 'ignored';
         }
     }
 
@@ -611,6 +640,7 @@ class WP_Object_Cache {
             'timeout' => 1,
             'read_timeout' => 1,
             'retry_interval' => null,
+            'persistent' => false,
         ];
 
         $settings = [
@@ -657,9 +687,10 @@ class WP_Object_Cache {
             $this->diagnostics[ 'shards' ] = WP_REDIS_SHARDS;
         } elseif ( defined( 'WP_REDIS_CLUSTER' ) ) {
             $args = [
-                'cluster' => array_values( WP_REDIS_CLUSTER ),
+                'cluster' => $this->build_cluster_connection_array(),
                 'timeout' => $parameters['timeout'],
                 'read_timeout' => $parameters['read_timeout'],
+                'persistent' => $parameters['persistent'],
             ];
 
             if ( isset( $parameters['password'] ) && version_compare( $version, '4.3.0', '>=' ) ) {
@@ -690,7 +721,7 @@ class WP_Object_Cache {
 
             if ( strcasecmp( 'unix', $parameters['scheme'] ) === 0 ) {
                 $args['host'] = $parameters['path'];
-                $args['port'] = null;
+                $args['port'] = -1;
             }
 
             if ( version_compare( $version, '3.1.3', '>=' ) ) {
@@ -834,7 +865,7 @@ class WP_Object_Cache {
             $parameters['servers'] = $servers;
             $options['replication'] = true;
         } elseif ( defined( 'WP_REDIS_CLUSTER' ) ) {
-            $servers = WP_REDIS_CLUSTER;
+            $servers = $this->build_cluster_connection_array();
             $parameters['cluster'] = $servers;
             $options['cluster'] = 'redis';
         }
@@ -875,6 +906,8 @@ class WP_Object_Cache {
      * @return void
      */
     protected function connect_using_credis( $parameters ) {
+        _doing_it_wrong( __FUNCTION__ , 'Credis support will be removed in future versions.' , '2.0.26' );
+
         $client = 'Credis';
 
         $creds_path = sprintf(
@@ -940,19 +973,24 @@ class WP_Object_Cache {
                 // phpcs:ignore WordPress.WP.AlternativeFunctions.parse_url_parse_url
                 $url_components = parse_url( $connection_string );
 
-                parse_str( $url_components['query'], $add_params );
+                if ( isset( $url_components['query'] ) ) {
+                    parse_str( $url_components['query'], $add_params );
+                }
 
                 if ( ! $is_cluster && isset( $add_params['alias'] ) ) {
                     $add_params['master'] = 'master' === $add_params['alias'];
                 }
 
                 $add_params['host'] = $url_components['host'];
+                $add_params['port'] = $url_components['port'];
 
                 if ( ! isset( $add_params['alias'] ) ) {
                     $add_params['alias'] = "redis-$index";
                 }
 
                 $clients[ $index ] = array_merge( $parameters, $add_params );
+
+                unset($add_params);
             }
 
             $this->redis = new Credis_Cluster( $clients );
@@ -1007,6 +1045,8 @@ class WP_Object_Cache {
      * @return void
      */
     protected function connect_using_hhvm( $parameters ) {
+        _doing_it_wrong( __FUNCTION__ , 'HHVM support will be removed in future versions.' , '2.0.26' );
+
         $this->redis = new Redis();
 
         // Adjust host and port if the scheme is `unix`.
@@ -1062,7 +1102,11 @@ class WP_Object_Cache {
         }
 
         if ( defined( 'WP_REDIS_CLUSTER' ) ) {
-            $info = $this->redis->info( current( array_values( WP_REDIS_CLUSTER ) ) );
+            $connectionID = current( $this->build_cluster_connection_array() );
+
+            $info = $this->determine_client() === 'predis'
+                ? $this->redis->getClientFor( $connectionID )->info()
+                : $this->redis->info( $connectionID );
         } else {
             $info = $this->redis->info();
         }
@@ -1132,10 +1176,107 @@ class WP_Object_Cache {
             return array_combine( array_keys( $data ), array_fill( 0, count( $data ), false ) );
         }
 
+        if ( $this->redis_status() && method_exists( $this->redis, 'pipeline' ) ) {
+            return $this->add_multiple_at_once( $data, $group, $expire );
+        }
+
         $values = [];
 
         foreach ( $data as $key => $value ) {
             $values[ $key ] = $this->add( $key, $value, $group, $expire );
+        }
+
+        return $values;
+    }
+
+    /**
+     * Adds multiple values to the cache in one call.
+     *
+     * @param array  $data   Array of keys and values to be added.
+     * @param string $group  Optional. Where the cache contents are grouped.
+     * @param int    $expire Optional. When to expire the cache contents, in seconds.
+     *                       Default 0 (no expiration).
+     * @return bool[] Array of return values, grouped by key. Each value is either
+     *                true on success, or false if cache key and group already exist.
+     */
+    protected function add_multiple_at_once( array $data, $group = 'default', $expire = 0 )
+    {
+        $keys = array_keys( $data );
+        $values = array_combine( $keys, array_fill( 0, count( $keys ), true ) );
+        $group = $this->sanitize_key_part( $group );
+
+        if ( ! $this->is_ignored_group( $group ) ) {
+            $start_time = microtime( true );
+
+            $orig_exp = $expire;
+            $expire = $this->validate_expiration( $expire );
+            $tx = $this->redis->pipeline();
+            $traceKV = [];
+
+            foreach ( $data as $key => $value ) {
+                /** This action is documented in includes/object-cache.php */
+                $expire = apply_filters( 'redis_cache_expiration', $expire, $key, $group, $orig_exp );
+
+                $key = $this->sanitize_key_part( $key );
+                $derived_key = $this->fast_build_key( $key, $group );
+
+                $values[ $key ] = $derived_key || ! isset( $this->cache[ $derived_key ] );
+
+                $args = [ $derived_key, $this->maybe_serialize( $value ) ];
+
+                if ( $this->redis instanceof Predis\Client ) {
+                    $args[] = 'nx';
+
+                    if ( $expire ) {
+                        $args[] = 'ex';
+                        $args[] = $expire;
+                    }
+                } else {
+                    if ( $expire ) {
+                        $args[] = [ 'nx', 'ex' => $expire ];
+                    } else {
+                        $args[] = [ 'nx' ];
+                    }
+                }
+
+                $traceKV[ $key ] = [
+                    'value' => $value,
+                    'status' => self::TRACE_FLAG_WRITE,
+                ];
+
+                $tx->set( ...$args );
+            }
+
+            try {
+                $method = ( $this->redis instanceof Predis\Client ) ? 'execute' : 'exec';
+
+                $values = array_map( function ( $response ) {
+                    return (bool) $this->parse_redis_response( $response );
+                }, $tx->{$method}() );
+
+                if ( $values ) {
+                    $values = array_combine( $keys, $values );
+                }
+
+                $execute_time = microtime( true ) - $start_time;
+
+                if ( $this->trace_enabled ) {
+                    $this->trace_command( 'set', $group, $traceKV, microtime( true ) - $start_time );
+                }
+
+                $this->cache_calls++;
+                $this->cache_time += $execute_time;
+            } catch ( Exception $exception ) {
+                $this->handle_exception( $exception );
+
+                return array_combine( $keys, array_fill( 0, count( $keys ), false ) );
+            }
+        }
+
+        foreach ( $values as $key => $value ) {
+            if ( $value ) {
+                $this->add_to_internal_cache( $this->fast_build_key( $key, $group ) , $value );
+            }
         }
 
         return $values;
@@ -1170,19 +1311,21 @@ class WP_Object_Cache {
      * @return  bool                   Returns TRUE on success or FALSE on failure.
      */
     protected function add_or_replace( $add, $key, $value, $group = 'default', $expiration = 0 ) {
-        $cache_addition_suspended = function_exists( 'wp_suspend_cache_addition' )
-            ? wp_suspend_cache_addition()
-            : false;
+        $cache_addition_suspended = function_exists( 'wp_suspend_cache_addition' ) && wp_suspend_cache_addition();
 
         if ( $add && $cache_addition_suspended ) {
             return false;
         }
 
         $result = true;
-        $derived_key = $this->build_key( $key, $group );
+
+        $san_key = $this->sanitize_key_part( $key );
+        $san_group = $this->sanitize_key_part( $group );
+
+        $derived_key = $this->fast_build_key( $san_key, $san_group );
 
         // Save if group not excluded and redis is up.
-        if ( ! $this->is_ignored_group( $group ) && $this->redis_status() ) {
+        if ( ! $this->is_ignored_group( $san_group ) && $this->redis_status() ) {
             try {
                 $orig_exp = $expiration;
                 $expiration = $this->validate_expiration( $expiration );
@@ -1275,7 +1418,11 @@ class WP_Object_Cache {
      */
     public function delete( $key, $group = 'default' ) {
         $result = false;
-        $derived_key = $this->build_key( $key, $group );
+
+        $san_key = $this->sanitize_key_part( $key );
+        $san_group = $this->sanitize_key_part( $group );
+
+        $derived_key = $this->fast_build_key( $san_key, $san_group );
 
         if ( isset( $this->cache[ $derived_key ] ) ) {
             unset( $this->cache[ $derived_key ] );
@@ -1284,7 +1431,7 @@ class WP_Object_Cache {
 
         $start_time = microtime( true );
 
-        if ( $this->redis_status() && ! $this->is_ignored_group( $group ) ) {
+        if ( $this->redis_status() && ! $this->is_ignored_group( $san_group ) ) {
             try {
                 $result = $this->parse_redis_response( $this->redis->del( $derived_key ) );
             } catch ( Exception $exception ) {
@@ -1354,6 +1501,9 @@ class WP_Object_Cache {
      *                true on success, or false if the contents were not deleted.
      */
     protected function delete_multiple_at_once( array $keys, $group = 'default' ) {
+        $start_time = microtime( true );
+        $traceKV = [];
+
         if ( $this->is_ignored_group( $group ) ) {
             $results = [];
 
@@ -1362,7 +1512,18 @@ class WP_Object_Cache {
 
                 $results[ $key ] = isset( $this->cache[ $derived_key ] );
 
+                $traceKV[ $key ] = [
+                    'value' => $this->cache[ $derived_key ],
+                    'status' => self::TRACE_FLAG_DEL,
+                ];
+
                 unset( $this->cache[ $derived_key ] );
+            }
+
+            $execute_time = microtime( true ) - $start_time;
+
+            if ( $this->trace_enabled ) {
+                $this->trace_command( 'set', $group, $traceKV, $execute_time );
             }
 
             return $results;
@@ -1374,6 +1535,11 @@ class WP_Object_Cache {
             foreach ($keys as $key) {
                 $derived_key = $this->build_key( (string) $key, $group );
 
+                $traceKV[ $key ] = [
+                    'value' => $this->cache[ $derived_key ],
+                    'status' => self::TRACE_FLAG_DEL,
+                ];
+
                 $tx->del( $derived_key );
 
                 unset( $this->cache[ $derived_key ] );
@@ -1384,6 +1550,12 @@ class WP_Object_Cache {
             $results = array_map( function ( $response ) {
                 return (bool) $this->parse_redis_response( $response );
             }, $tx->{$method}() );
+
+            $execute_time = microtime( true ) - $start_time;
+
+            if ( $this->trace_enabled ) {
+                $this->trace_command( 'set', $group, $traceKV, $execute_time );
+            }
 
             return array_combine( $keys, $results );
         } catch ( Exception $exception ) {
@@ -1653,7 +1825,10 @@ LUA;
         }
 
         $start_time = microtime( true );
-        $derived_key = $this->build_key( $key, $group );
+
+        $san_key = $this->sanitize_key_part( $key );
+        $san_group = $this->sanitize_key_part( $group );
+        $derived_key = $this->fast_build_key( $san_key, $san_group );
 
         if ( isset( $this->cache[ $derived_key ] ) && ! $force ) {
             $found = true;
@@ -1670,7 +1845,7 @@ LUA;
             }
 
             return $value;
-        } elseif ( $this->is_ignored_group( $group ) || ! $this->redis_status() ) {
+        } elseif ( $this->is_ignored_group( $san_group ) || ! $this->redis_status() ) {
             $found = false;
             $this->cache_misses++;
 
@@ -1789,11 +1964,14 @@ LUA;
         $derived_keys = [];
         $start_time = microtime( true );
 
+        $san_group = $this->sanitize_key_part( $group );
+
         foreach ( $keys as $key ) {
-            $derived_keys[ $key ] = $this->build_key( $key, $group );
+            $san_key = $this->sanitize_key_part( $key );
+            $derived_keys[ $key ] = $this->fast_build_key( $san_key, $san_group );
         }
 
-        if ( $this->is_ignored_group( $group ) || ! $this->redis_status() ) {
+        if ( $this->is_ignored_group( $san_group ) || ! $this->redis_status() ) {
             $traceKV = [];
 
             foreach ( $keys as $key ) {
@@ -1971,10 +2149,14 @@ LUA;
     public function set( $key, $value, $group = 'default', $expiration = 0 ) {
         $result = true;
         $start_time = microtime( true );
-        $derived_key = $this->build_key( $key, $group );
+
+        $san_key = $this->sanitize_key_part( $key );
+        $san_group = $this->sanitize_key_part( $group );
+
+        $derived_key = $this->fast_build_key( $san_key, $san_group );
 
         // Save if group not excluded from redis and redis is up.
-        if ( ! $this->is_ignored_group( $group ) && $this->redis_status() ) {
+        if ( ! $this->is_ignored_group( $san_group ) && $this->redis_status() ) {
             $orig_exp = $expiration;
             $expiration = $this->validate_expiration( $expiration );
 
@@ -2048,11 +2230,107 @@ LUA;
      *                       Default 0 (no expiration).
      * @return bool[] Array of return values, grouped by key. Each value is always true.
      */
-    public function set_multiple( array $data, $group = 'default', $expiration = 0 ) {
+    public function set_multiple( array $data, $group = 'default', $expire = 0 ) {
+        if ( $this->redis_status() && method_exists( $this->redis, 'pipeline' ) ) {
+            return $this->set_multiple_at_once( $data, $group, $expire );
+        }
+
         $values = [];
 
         foreach ( $data as $key => $value ) {
-            $values[ $key ] = $this->set( $key, $value, $group, $expiration );
+            $values[ $key ] = $this->set( $key, $value, $group, $expire );
+        }
+
+        return $values;
+    }
+
+    /**
+     * Sets multiple values to the cache in one call.
+     *
+     * @param array  $data       Array of key and value to be set.
+     * @param string $group      Optional. Where the cache contents are grouped.
+     * @param int    $expiration Optional. When to expire the cache contents, in seconds.
+     *                           Default 0 (no expiration).
+     * @return bool[] Array of return values, grouped by key. Each value is always true.
+     */
+    protected function set_multiple_at_once( array $data, $group = 'default', $expiration = 0 )
+    {
+        $values = [];
+        $group = $this->sanitize_key_part( $group );
+
+        if ( ! $this->is_ignored_group( $group ) ) {
+            $start_time = microtime( true );
+
+            $orig_exp = $expiration;
+            $expiration = $this->validate_expiration( $expiration );
+            $tx = $this->redis->pipeline();
+            $keys = array_keys( $data );
+            $traceKV = [];
+
+            foreach ( $data as $key => $value ) {
+                $key = $this->sanitize_key_part( $key );
+                $derived_key = $this->fast_build_key( $key, $group );
+
+                /** This action is documented in includes/object-cache.php */
+                $expiration = apply_filters( 'redis_cache_expiration', $expiration, $key, $group, $orig_exp );
+
+                $args = [ $derived_key, $this->maybe_serialize( $value ) ];
+
+                if ( $this->redis instanceof Predis\Client ) {
+                    $args[] = 'nx';
+
+                    if ( $expiration ) {
+                        $args[] = 'ex';
+                        $args[] = $expiration;
+                    }
+                } else {
+                    if ( $expiration ) {
+                        $args[] = [ 'nx', 'ex' => $expiration ];
+                    } else {
+                        $args[] = [ 'nx' ];
+                    }
+                }
+
+                $traceKV[ $key ] = [
+                    'value' => $value,
+                    'status' => self::TRACE_FLAG_WRITE,
+                ];
+
+                $tx->set( ...$args );
+            }
+
+            try {
+                $method = ( $this->redis instanceof Predis\Client ) ? 'execute' : 'exec';
+
+                $values = array_map( function ( $response ) {
+                    return (bool) $this->parse_redis_response( $response );
+                }, $tx->{$method}() );
+
+                if ( $values ) {
+                    $values = array_combine( $keys, $values );
+                }
+            } catch ( Exception $exception ) {
+                $this->handle_exception( $exception );
+
+                return array_combine( $keys, array_fill( 0, count( $keys ), false ) );
+            }
+
+            $execute_time = microtime( true ) - $start_time;
+            $this->cache_calls++;
+            $this->cache_time += $execute_time;
+
+            if ( $this->trace_enabled ) {
+                $this->trace_command( 'set', $group, $traceKV, $execute_time );
+            }
+        }
+
+        foreach ($data as $key => $value) {
+            $key = $this->sanitize_key_part( $key );
+            $derived_key = $this->fast_build_key( $key, $group );
+
+            if ( isset( $values[ $key ] ) && $values[ $key ] ) {
+                $this->add_to_internal_cache( $derived_key, $value );
+            }
         }
 
         return $values;
@@ -2069,11 +2347,15 @@ LUA;
     public function increment( $key, $offset = 1, $group = 'default' ) {
         $offset = (int) $offset;
         $start_time = microtime( true );
-        $derived_key = $this->build_key( $key, $group );
+
+        $san_key = $this->sanitize_key_part( $key );
+        $san_group = $this->sanitize_key_part( $group );
+
+        $derived_key = $this->fast_build_key( $san_key, $san_group );
         $trace_flags = self::TRACE_FLAG_READ | self::TRACE_FLAG_WRITE;
 
         // If group is a non-Redis group, save to internal cache, not Redis.
-        if ( $this->is_ignored_group( $group ) || ! $this->redis_status() ) {
+        if ( $this->is_ignored_group( $san_group ) || ! $this->redis_status() ) {
             $value = $this->get_from_internal_cache( $derived_key );
             $value += $offset;
             $this->add_to_internal_cache( $derived_key, $value );
@@ -2141,11 +2423,15 @@ LUA;
     public function decrement( $key, $offset = 1, $group = 'default' ) {
         $offset = (int) $offset;
         $start_time = microtime( true );
-        $derived_key = $this->build_key( $key, $group );
+
+        $san_key = $this->sanitize_key_part( $key );
+        $san_group = $this->sanitize_key_part( $group );
+
+        $derived_key = $this->fast_build_key( $san_key, $san_group );
         $trace_flags = self::TRACE_FLAG_READ | self::TRACE_FLAG_WRITE;
 
         // If group is a non-Redis group, save to internal cache, not Redis.
-        if ( $this->is_ignored_group( $group ) || ! $this->redis_status() ) {
+        if ( $this->is_ignored_group( $san_group ) || ! $this->redis_status() ) {
             $value = $this->get_from_internal_cache( $derived_key );
             $value -= $offset;
             $this->add_to_internal_cache( $derived_key, $value );
@@ -2261,7 +2547,7 @@ LUA;
             ],
             'errors' => empty( $this->errors ) ? null : $this->errors,
             'meta' => [
-                'Client' => $this->diagnostics['client'] ?: 'Unknown',
+                'Client' => $this->diagnostics['client'] ?? 'Unknown',
                 'Redis Version' => $this->redis_version,
             ],
         ];
@@ -2270,8 +2556,8 @@ LUA;
     /**
      * Builds a key for the cached object using the prefix, group and key.
      *
-     * @param   string $key        The key under which to store the value.
-     * @param   string $group      The group value appended to the $key.
+     * @param   string $key        The key under which to store the value, pre-sanitized.
+     * @param   string $group      The group value appended to the $key, pre-sanitized.
      *
      * @return  string
      */
@@ -2280,11 +2566,29 @@ LUA;
             $group = 'default';
         }
 
+        $san_key = $this->sanitize_key_part( $key );
+        $san_group = $this->sanitize_key_part( $group );
+
+        return $this->fast_build_key($san_key, $san_group);
+    }
+
+    /**
+     * Builds a key for the cached object using the prefix, group and key.
+     *
+     * @param   string $key        The key under which to store the value, pre-sanitized.
+     * @param   string $group      The group value appended to the $key, pre-sanitized.
+     *
+     * @return  string
+     */
+    public function fast_build_key( $key, $group = 'default' ) {
+        if ( empty( $group ) ) {
+            $group = 'default';
+        }
+
         $salt = defined( 'WP_REDIS_PREFIX' ) ? trim( WP_REDIS_PREFIX ) : '';
+
         $prefix = $this->is_global_group( $group ) ? $this->global_prefix : $this->blog_prefix;
 
-        $key = $this->sanitize_key_part( $key );
-        $group = $this->sanitize_key_part( $group );
 
         $prefix = trim( $prefix, '_-:$' );
 
@@ -2304,31 +2608,43 @@ LUA;
     /**
      * Checks if the given group is part the ignored group array
      *
-     * @param string $group  Name of the group to check.
+     * @param string $group  Name of the group to check, pre-sanitized.
      * @return bool
      */
     protected function is_ignored_group( $group ) {
-        return in_array( $this->sanitize_key_part( $group ), $this->ignored_groups, true );
+        return $this->is_group_of_type( $group, 'ignored' );
     }
 
     /**
      * Checks if the given group is part the global group array
      *
-     * @param string $group  Name of the group to check.
+     * @param string $group  Name of the group to check, pre-sanitized.
      * @return bool
      */
     protected function is_global_group( $group ) {
-        return in_array( $this->sanitize_key_part( $group ), $this->global_groups, true );
+        return $this->is_group_of_type( $group, 'global' );
     }
 
     /**
      * Checks if the given group is part the unflushable group array
      *
-     * @param string $group  Name of the group to check.
+     * @param string $group  Name of the group to check, pre-sanitized.
      * @return bool
      */
     protected function is_unflushable_group( $group ) {
-        return in_array( $this->sanitize_key_part( $group ), $this->unflushable_groups, true );
+        return $this->is_group_of_type( $group, 'unflushable' );
+    }
+
+    /**
+     * Checks the type of the given group
+     *
+     * @param string $group  Name of the group to check, pre-sanitized.
+     * @param string $type   Type of the group to check.
+     * @return bool
+     */
+    private function is_group_of_type( $group, $type ) {
+        return isset( $this->group_type[ $group ] )
+            && $this->group_type[ $group ] == $type;
     }
 
     /**
@@ -2415,6 +2731,8 @@ LUA;
         } else {
             $this->ignored_groups = array_unique( array_merge( $this->ignored_groups, $groups ) );
         }
+
+        $this->cache_group_types();
     }
 
     /**
@@ -2426,6 +2744,7 @@ LUA;
         $groups = (array) $groups;
 
         $this->ignored_groups = array_unique( array_merge( $this->ignored_groups, $groups ) );
+        $this->cache_group_types();
     }
 
     /**
@@ -2437,6 +2756,7 @@ LUA;
         $groups = (array) $groups;
 
         $this->unflushable_groups = array_unique( array_merge( $this->unflushable_groups, $groups ) );
+        $this->cache_group_types();
     }
 
     /**
@@ -2661,6 +2981,27 @@ LUA;
              */
             do_action( 'redis_object_cache_error', $exception );
         }
+    }
+
+    /**
+     * Builds a clean connection array out of redis clusters array.
+     *
+     * @return  array
+     */
+    protected function build_cluster_connection_array() {
+        $cluster = array_values( WP_REDIS_CLUSTER );
+
+        foreach ( $cluster as $key => $server ) {
+            $connection_string = parse_url( $server );
+
+            $cluster[ $key ] = sprintf(
+                "%s:%s",
+                $connection_string['host'],
+                $connection_string['port']
+            );
+        }
+
+        return $cluster;
     }
 
     /**